+++ docs/app/pages/docs/[...slug]/index.vue
@@ -153,7 +153,7 @@ onMounted(() => {
 <template>
   <B24DashboardPanel
     id="doc"
-    :b24ui="{ body: 'items-stretch justify-between scrollbar-transparent' }"
+    :b24ui="{ body: 'items-stretch justify-between scrollbar-transparent scrollbar-both-edges' }"
   >
     <template #header>
       <Header />
@@ -162,7 +162,7 @@ onMounted(() => {
     <template v-if="page" #body>
       <B24Page
         :b24ui="{
-          center: 'flex flex-col gap-6',
+          center: 'flex flex-col lg:gap-2.5',
           right: 'lg:col-span-2 order-first lg:order-last lg:top-[0px]'
         }"
       >
@@ -230,7 +230,7 @@ onMounted(() => {
             <B24Card
               variant="outline-alt"
               class="rounded-none lg:rounded-(--ui-border-radius-md) backdrop-blur-md border-0"
-              :b24ui="{ body: 'px-[22px] py-0 sm:ps-[22px] lg:ps-[10px] lg:pe-[4px] sm:py-0 pt-[12px] sm:pt-[12px] lg:py-[15px]' }"
+              :b24ui="{ body: 'px-3 py-0 sm:ps-[22px] lg:ps-[10px] lg:pe-1 sm:py-0 pt-3 sm:pt-3 lg:py-[15px]' }"
             >
               <B24ContentToc
                 :links="page.body.toc.links"
@@ -244,6 +244,10 @@ onMounted(() => {
           v-if="page"
           as="main"
           :class="cardColorContext"
+          :b24ui="{
+            root: 'rounded-none lg:rounded-(--ui-border-radius-md)',
+            body: 'p-3'
+          }"
         >
           <ContentRenderer v-if="page.body" :value="page" />